adding router + home page
