+++ src/Components/Home.css
@@ -1,0 +1,23 @@
+.input {
+  width: 90%;
+  margin-bottom: 3%;
+  display: flex;
+}
+.title {
+  display: flex;
+  justify-content: center;
+  margin-top: 3%;
+  margin-bottom: 3%;
+  font-size: 30px;
+  font-family: 'Lucida Sans', 'Lucida Sans Regular', 'Lucida Grande', 'Lucida Sans Unicode', Geneva, Verdana, sans-serif;
+  color: rgb(57, 4, 182);
+}
+.image {
+  display: flex;
+  justify-content: center;
+}
+.checkbox {
+  display: flex;
+  justify-content: center;
+  size: 10%;
+}
