+++ src/Components/SearchPoint.jsx
@@ -1,0 +1,73 @@
+import React, { Component } from 'react';
+import TopBar from './topBar';
+import Content from './content';
+import BottomBar from './bottomBar';
+import getDataFromReddit from '../APIFunctions/getDataFromReddit';
+
+class SearchPoint extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      keyWord: '',
+      redditData: [],
+      twitterData: [],
+      hackerNoonData: [],
+      sourceToggle: {
+        twitter: true,
+        reddit: true,
+        hacker: true,
+      },
+    };
+    this.refreshDataFromAPI = this.refreshDataFromAPI.bind(this);
+  }
+
+  refreshDataFromAPI(keyWord = 'Apple', sort = 'relevance', nbOfItems = 25) {
+    // Refresh data from Reddit
+    getDataFromReddit(keyWord, sort, nbOfItems).then((result) => this.setState({ redditData: result }));
+
+    // refresh data from Twitter
+    // Add the code here...
+
+    // refresh data from Hacker noon
+    // Add the code here...
+  }
+
+  handleChange = (event) => {
+    this.setState({
+      keyWord: event.target.value,
+    });
+  };
+
+  handleClick = (event) => {
+    this.refreshDataFromAPI(this.state.keyWord);
+  };
+
+  handleToggle = (source) => {
+    let sourceToggle = { ...this.state.sourceToggle };
+    sourceToggle[source] = !sourceToggle[source];
+    this.setState({ sourceToggle });
+  };
+
+  render() {
+    return (
+      <div className="App">
+        <TopBar
+          toggle={this.state.sourceToggle}
+          onToggle={this.handleToggle}
+          changeHandler={this.handleChange}
+          clickHandler={this.handleClick}
+          searchValue={this.state.keyWord}
+        />
+        <Content
+          toggle={this.state.sourceToggle}
+          redditData={this.state.redditData}
+          twitterData={this.state.twitterData}
+          hackerNoonData={this.state.hackerNoonData}
+        />
+        <BottomBar />
+      </div>
+    );
+  }
+}
+
+export default SearchPoint;
